refactor: variable header 구조 변경
각 패킷 타입의 variable header를 패킷 클래스 안에서 정의하도록 변경경
