+++ broker/packet/mqtt_control_packet.cpp
@@ -2,9 +2,9 @@
 #include <iostream>
 #include "mqtt_control_packet.h"
 #include "fixed_header.h"
-#include "variable_header.h"
 #include "connect_packet.h"
 #include "connack_packet.h"
+#include "publish_packet.h"
 
 // Parse byte stream to a packet format.
 std::unique_ptr<mqtt_control_packet> mqtt_control_packet::mqtt_control_packet::parse(const uint8_t* data, size_t size)
@@ -16,7 +16,7 @@ std::unique_ptr<mqtt_control_packet> mqtt_control_packet::mqtt_control_packet::p
     fixed_header header = fixed_header::parse(data, size);
     
     // Debug fixed hedaer.
-    header.print_fixed_header();
+    header.debug();
 
     // Payload pointer. (or variable header pointer to some packet types.)
     const uint8_t* payload = data + header.header_length;
@@ -27,6 +27,8 @@ std::unique_ptr<mqtt_control_packet> mqtt_control_packet::mqtt_control_packet::p
             return connect_packet::parse(payload, header.remaining_length);
         case mqtt_packet_type::CONNACK:
             return NULL;
+        case mqtt_packet_type::PUBLISH:
+            return publish_packet::parse(payload, header.remaining_length);
         // Other parse functions...
         default:
             throw std::runtime_error("Unsupported packet type");
